test(acp): isolate lifecycle history fixtures
